Centralize device entity setup in a base entity
Add MelCloudDescriptionEntity holding the unique_id and device-info setup that
the binary sensor and sensor platforms duplicate, and have them inherit from it.
Non-functional change.
